+++ app/src/main/java/com/example/uistatewithflowtest/repository/ScrapRepository.kt
@@ -3,7 +3,7 @@ package com.example.uistatewithflowtest.repository
 import kotlinx.coroutines.delay
 import javax.inject.Singleton
 
-data class Scrap(val value: Int) {
+data class Scrap(val value: Long) {
 
     override fun toString(): String {
         return "Scrap($value)"
@@ -15,7 +15,7 @@ class ScrapRepository(
     private val delayBy: Long = 1000
 ) {
 
-    suspend fun get(id: Int): Scrap? {
+    suspend fun get(id: Long): Scrap? {
         delay(delayBy)
         return if ((0 until 4).random() == 0) Scrap(id)  else null
     }
